refactor: extract Model / Layout / Aggregation into compilertop package
Continues the breakup of CompilerTop.scala. Three new files:

  - Model.scala       — PhaseFilter, Sort, ModuleStats, and the phase-name
                        classification sets (FrontendPhases, MonoCountPhases,
                        OptCountPhases, ClsCountPhases).
  - Layout.scala      — Layout case class + companion `compute()` (idiomatic
                        class+companion; the algorithm lives next to the
                        data type it produces).
  - Aggregation.scala — matchesFilter, applyFilter, sumPhaseCounts,
                        defSortKey, moduleSortKey, aggregateByModule.

Pure code move, no behavior change. `applyFilter` moved off `class
CompilerTop` after confirming it reads no class state. CompilerTop.scala
shrinks by ~280 lines and is now mostly lifecycle + rendering.
